# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Interview Position Tracker API",
    "description": "\n    A comprehensive REST API for tracking job positions and interview progress.\n    \n    ## Features\n    \n    * **User Authentication**: Secure JWT-based authentication system\n    * **Position Management**: Create, read, update, and delete job positions\n    * **Interview Tracking**: Track multiple interview stages for each position\n    * **Statistics & Analytics**: Get insights into your job search progress\n    * **Data Filtering**: Filter positions by status, company, date range, and more\n    * **Secure Access**: User-specific data isolation and authorization\n    \n    ## Authentication\n    \n    Most endpoints require authentication. To authenticate:\n    \n    1. Register a new account using `/api/v1/auth/register`\n    2. Login using `/api/v1/auth/login` to get an access token\n    3. Include the token in the Authorization header: `Bearer <your_token>`\n    \n    ## Workflow\n    \n    1. **Register/Login** to get authenticated\n    2. **Create positions** for jobs you're applying to\n    3. **Add interviews** as you progress through hiring processes\n    4. **Update statuses** as your applications move forward\n    5. **View statistics** to analyze your job search performance\n    ",
    "contact": {
      "name": "Interview Position Tracker API",
      "url": "https://github.com/your-repo/interview-tracker"
    },
    "license": {
      "name": "MIT",
...
  },
  "paths": {
...
    "/api/v1/positions/{position_id}/interviews": {
      "post": {
        "tags": [
          "Interviews"
        ],
        "summary": "Create Interview",
        "description": "Create a new interview for a position.\n\nCreates a new interview record for the specified position if it exists and belongs to the authenticated user.",
        "operationId": "create_interview_api_v1_positions__position_id__interviews_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Position Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Interviews"
        ],
        "summary": "List Interviews",
        "description": "List all interviews for a position.\n\nReturns all interview records for the specified position if it exists and belongs to the authenticated user.\nInterviews are ordered by scheduled date.",
        "operationId": "list_interviews_api_v1_positions__position_id__interviews_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Position Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/{interview_id}": {
      "get": {
        "tags": [
          "Interviews"
        ],
        "summary": "Get Interview",
        "description": "Get a specific interview by ID.\n\nReturns the interview details if it exists and its associated position belongs to the authenticated user.",
        "operationId": "get_interview_api_v1_interviews__interview_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Interviews"
        ],
        "summary": "Update Interview",
        "description": "Update a specific interview.\n\nUpdates the interview with the provided data if it exists and its associated position belongs to the authenticated user.\nOnly provided fields will be updated.",
        "operationId": "update_interview_api_v1_interviews__interview_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interview Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Interviews"
        ],
        "summary": "Delete Interview",
        "description": "Delete a specific interview.\n\nDeletes the interview if it exists and its associated position belongs to the authenticated user.",
        "operationId": "delete_interview_api_v1_interviews__interview_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interview Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/{interview_id}/schedule": {
      "put": {
        "tags": [
          "Interviews"
        ],
        "summary": "Update Interview Schedule",
        "description": "Update the scheduled date of a specific interview.\n\nUpdates only the scheduled date of the interview if it exists and its associated position belongs to the authenticated user.",
        "operationId": "update_interview_schedule_api_v1_interviews__interview_id__schedule_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interview Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewScheduleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/{interview_id}/notes": {
      "put": {
        "tags": [
          "Interviews"
        ],
        "summary": "Update Interview Notes",
        "description": "Update the notes of a specific interview.\n\nUpdates only the notes of the interview if it exists and its associated position belongs to the authenticated user.",
        "operationId": "update_interview_notes_api_v1_interviews__interview_id__notes_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interview Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewNotesUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/{interview_id}/outcome": {
      "put": {
        "tags": [
          "Interviews"
        ],
        "summary": "Update Interview Outcome",
        "description": "Update the outcome of a specific interview.\n\nUpdates only the outcome of the interview if it exists and its associated position belongs to the authenticated user.\nIf the outcome is 'failed', the associated position status will be updated to 'rejected'.",
        "operationId": "update_interview_outcome_api_v1_interviews__interview_id__outcome_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interview Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewOutcomeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterviewCreate": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/InterviewType",
            "description": "Type of interview"
          },
          "place": {
            "$ref": "#/components/schemas/InterviewPlace",
            "description": "Format/location of interview"
          },
          "scheduled_date": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled Date",
            "description": "Scheduled date and time of interview"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 480.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes",
            "description": "Interview duration in minutes"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Additional notes about the interview"
          },
          "outcome": {
            "$ref": "#/components/schemas/InterviewOutcome",
            "description": "Interview outcome",
            "default": "pending"
          }
        },
        "type": "object",
        "required": [
          "type",
          "place",
          "scheduled_date"
        ],
        "title": "InterviewCreate",
        "description": "Schema for creating a new interview."
      },
      "InterviewListResponse": {
        "properties": {
          "interviews": {
            "items": {
              "$ref": "#/components/schemas/InterviewResponse"
            },
            "type": "array",
            "title": "Interviews",
            "description": "List of interviews"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of interviews"
          }
        },
        "type": "object",
        "required": [
          "interviews",
          "total"
        ],
        "title": "InterviewListResponse",
        "description": "Schema for interview list response."
      },
      "InterviewNotesUpdate": {
        "properties": {
          "notes": {
            "type": "string",
            "title": "Notes",
            "description": "Updated notes about the interview"
          }
        },
        "type": "object",
        "required": [
          "notes"
        ],
        "title": "InterviewNotesUpdate",
        "description": "Schema for updating interview notes only."
      },
      "InterviewOutcome": {
        "type": "string",
        "enum": [
          "pending",
          "passed",
          "failed",
          "cancelled"
        ],
        "title": "InterviewOutcome",
        "description": "Outcome of an interview."
      },
      "InterviewOutcomeUpdate": {
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/InterviewOutcome",
            "description": "Interview outcome"
          }
        },
        "type": "object",
        "required": [
          "outcome"
        ],
        "title": "InterviewOutcomeUpdate",
        "description": "Schema for updating interview outcome only."
      },
      "InterviewPlace": {
        "type": "string",
        "enum": [
          "phone",
          "video",
          "onsite"
        ],
        "title": "InterviewPlace",
        "description": "Format/location where the interview takes place."
      },
      "InterviewResponse": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/InterviewType",
            "description": "Type of interview"
          },
          "place": {
            "$ref": "#/components/schemas/InterviewPlace",
            "description": "Format/location of interview"
          },
          "scheduled_date": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled Date",
            "description": "Scheduled date and time of interview"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 480.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes",
            "description": "Interview duration in minutes"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Additional notes about the interview"
          },
          "outcome": {
            "$ref": "#/components/schemas/InterviewOutcome",
            "description": "Interview outcome",
            "default": "pending"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Interview unique identifier"
          },
          "position_id": {
            "type": "string",
            "format": "uuid",
            "title": "Position Id",
            "description": "Associated position identifier"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Interview creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Interview last update timestamp"
          }
        },
        "type": "object",
        "required": [
          "type",
          "place",
          "scheduled_date",
          "id",
          "position_id",
          "created_at",
          "updated_at"
        ],
        "title": "InterviewResponse",
        "description": "Schema for interview response data."
      },
      "InterviewScheduleUpdate": {
        "properties": {
          "scheduled_date": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled Date",
            "description": "New scheduled date and time of interview"
          }
        },
        "type": "object",
        "required": [
          "scheduled_date"
        ],
        "title": "InterviewScheduleUpdate",
        "description": "Schema for updating interview scheduled date only."
      },
      "InterviewType": {
        "type": "string",
        "enum": [
          "technical",
          "behavioral",
          "hr",
          "final"
        ],
        "title": "InterviewType",
        "description": "Type of interview based on its purpose."
      },
      "InterviewUpdate": {
        "properties": {
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterviewType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Type of interview"
          },
          "place": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterviewPlace"
              },
              {
                "type": "null"
              }
            ],
            "description": "Format/location of interview"
          },
          "scheduled_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Date",
            "description": "Scheduled date and time of interview"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 480.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes",
            "description": "Interview duration in minutes"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Additional notes about the interview"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterviewOutcome"
              },
              {
                "type": "null"
              }
            ],
            "description": "Interview outcome"
          }
        },
        "type": "object",
        "title": "InterviewUpdate",
        "description": "Schema for updating an existing interview."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}